Reorganise: Subsume inlinekeyboards package into new assemblers package
