+++ src/internal/assemblers/inline_keyboards.go
@@ -1,12 +1,12 @@
-package inlinekeyboards
+package assemblers
 
 import (
 	tgbotapi "github.com/go-telegram-bot-api/telegram-bot-api/v5"
 	"github.com/jonleeyz/bball8bot/internal/handlers/callbacks"
 )
 
-// BuildInlineKeyboard builds a basic inline keyboard for the training poll template messsage.
-func BuildTrainingPollInlineKeyboard() tgbotapi.InlineKeyboardMarkup {
+// AssembleTrainingPollInlineKeyboard builds a basic inline keyboard for the training poll template messsage.
+func AssembleTrainingPollInlineKeyboard() tgbotapi.InlineKeyboardMarkup {
 	return tgbotapi.NewInlineKeyboardMarkup(
 		tgbotapi.NewInlineKeyboardRow(
 			tgbotapi.NewInlineKeyboardButtonData("Attending", callbacks.CALLBACK_DATA_ATTENDING),
